ref(tiller): refactor tiller APIs into logical files
*ReleaseServer.RollbackRelease -> tiller/release_rollback.go
